Exceptions/Backend/Backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-crispy-forms==2.3
crispy-tailwind==1.0.3
crispy-bootstrap4==2023.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-qr-code==3.1.1
pillow==11.1.0
sqlparse==0.5.3
